requirements.txt
paho-mqtt
customtkinter